rename 'store' to 'storage' to improve semantics
