+++ pkg/storage/doc.go
@@ -1,0 +1,3 @@
+// Package storage contains interface implemented by rqlite cluster
+// and data model used by bpmn & dmn engines
+package storage
