+++ pkg/storage/storage.go
@@ -1,4 +1,4 @@
-package store
+package storage
 
 import (
 	"context"
