--- pkg/store/doc.go
@@ -1,3 +1,0 @@
-// Package contains store interface implemented by rqlite cluster
-// and data model used by bpmn & dmn engines
-package store
